Add image processing service, russian error localization and tests

+++ admin_service/src/main/java/com/hse/adminservice/operations/queue/dto/ServiceRequestQueueItemResponse.java
@@ -12,7 +12,7 @@ public record ServiceRequestQueueItemResponse(
         String userName,
         String typeName,
         String name,
-        Integer cost,
+        Long cost,
         String status,
         LocalDate createdAt
 ) {
